# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Assessment — Channel & Upstream",
    "version": "1.0.0",
    "description": "Merged OpenAPI 3.0 document for the demo NestJS + Fastify services."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Channel service (override port via CHANNEL_PORT)"
    },
    {
      "url": "http://127.0.0.1:3001",
      "description": "Upstream simulator (override via UPSTREAM_PORT)"
    }
  ],
  "paths": {
    "/channel/health": {
      "get": {
        "operationId": "ChannelController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "channel"
        ]
      }
    },
    "/channel/integration/status": {
      "get": {
        "operationId": "ChannelController_integrationStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Expose upstream circuit breaker state (opossum)",
        "tags": [
          "channel"
        ]
      }
    },
    "/channel/demo/call": {
      "post": {
        "operationId": "ChannelController_demoCall",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "traceparent",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Invoke flaky upstream through integration framework (CQRS)",
        "tags": [
          "channel"
        ]
      }
    },
    "/upstream/health": {
      "get": {
        "operationId": "UpstreamController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "upstream"
        ]
      }
    },
    "/upstream/simulate/config": {
      "get": {
        "operationId": "UpstreamController_simulateConfig",
        "parameters": [
          {
            "name": "failureRate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Tune failure rate at runtime (CQRS command)",
        "tags": [
          "upstream"
        ]
      }
    },
    "/upstream/echo": {
      "post": {
        "operationId": "UpstreamController_echo",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "hello"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Echo payload via CQRS command + flaky upstream simulation",
        "tags": [
          "upstream"
        ]
      }
    }
  },
  "components": {
    "schemas": {}
  },
  "tags": []
}